ironclad/requirements.txt
torch
torchvision
facenet-pytorch
Pillow
faiss-cpu
flask
numpy
pandas